requirements.txt
requests
websocket-client
rich
python-dateutil
certifi
streamlit
plotly
pandas
